chore: extract `ConsoleMessage` and `FileChooser` into its own module
