+++ src/ConsoleMessage.ts
@@ -1,0 +1,58 @@
+/**
+ * Copyright 2017 Google Inc. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import { JSHandle } from './JSHandle';
+
+interface ConsoleMessageLocation {
+  url?: string;
+  lineNumber?: number;
+  columnNumber?: number;
+}
+
+export class ConsoleMessage {
+  _type: string;
+  _text: string;
+  _args: JSHandle[];
+  _location: ConsoleMessageLocation;
+
+  constructor(
+    type: string,
+    text: string,
+    args: JSHandle[],
+    location: ConsoleMessageLocation = {}
+  ) {
+    this._type = type;
+    this._text = text;
+    this._args = args;
+    this._location = location;
+  }
+
+  type(): string {
+    return this._type;
+  }
+
+  text(): string {
+    return this._text;
+  }
+
+  args(): JSHandle[] {
+    return this._args;
+  }
+
+  location(): ConsoleMessageLocation {
+    return this._location;
+  }
+}
