+++ src/Page.ts
@@ -38,80 +38,14 @@ import {
 } from './NetworkManager';
 import { Accessibility } from './Accessibility';
 import { TimeoutSettings } from './TimeoutSettings';
-import { PuppeteerLifeCycleEvent } from './LifecycleWatcher';
+import { FileChooser } from './FileChooser';
+import { ConsoleMessage } from './ConsoleMessage'
+import {
+  ScreenshotOptions, PaperFormat, Metrics, WaitForOptions, MediaFeature,
+  ScreenshotClip, PDFOptions
+} from './types'
 
 const writeFileAsync = helper.promisify(fs.writeFile);
-
-interface Metrics {
-  Timestamp?: number;
-  Documents?: number;
-  Frames?: number;
-  JSEventListeners?: number;
-  Nodes?: number;
-  LayoutCount?: number;
-  RecalcStyleCount?: number;
-  LayoutDuration?: number;
-  RecalcStyleDuration?: number;
-  ScriptDuration?: number;
-  TaskDuration?: number;
-  JSHeapUsedSize?: number;
-  JSHeapTotalSize?: number;
-}
-
-interface WaitForOptions {
-  timeout?: number;
-  waitUntil?: PuppeteerLifeCycleEvent | PuppeteerLifeCycleEvent[];
-}
-
-interface MediaFeature {
-  name: string;
-  value: string;
-}
-
-interface ScreenshotClip {
-  x: number;
-  y: number;
-  width: number;
-  height: number;
-}
-
-interface ScreenshotOptions {
-  type?: 'png' | 'jpeg';
-  path?: string;
-  fullPage?: boolean;
-  clip?: ScreenshotClip;
-  quality?: number;
-  omitBackground?: boolean;
-  encoding?: string;
-}
-
-interface PDFMargin {
-  top?: string | number;
-  bottom?: string | number;
-  left?: string | number;
-  right?: string | number;
-}
-
-interface PDFOptions {
-  scale?: number;
-  displayHeaderFooter?: boolean;
-  headerTemplate?: string;
-  footerTemplate?: string;
-  printBackground?: boolean;
-  landscape?: boolean;
-  pageRanges?: string;
-  format?: string;
-  width?: string | number;
-  height?: string | number;
-  preferCSSPageSize?: boolean;
-  margin?: PDFMargin;
-  path?: string;
-}
-
-interface PaperFormat {
-  width: number;
-  height: number;
-}
 
 const paperFormats: Record<string, PaperFormat> = {
   letter: { width: 8.5, height: 11 },
@@ -1318,83 +1252,4 @@ function convertPrintParameterToInches(
     );
   }
   return pixels / 96;
-}
-
-interface ConsoleMessageLocation {
-  url?: string;
-  lineNumber?: number;
-  columnNumber?: number;
-}
-
-export class ConsoleMessage {
-  _type: string;
-  _text: string;
-  _args: JSHandle[];
-  _location: ConsoleMessageLocation;
-
-  constructor(
-    type: string,
-    text: string,
-    args: JSHandle[],
-    location: ConsoleMessageLocation = {}
-  ) {
-    this._type = type;
-    this._text = text;
-    this._args = args;
-    this._location = location;
-  }
-
-  type(): string {
-    return this._type;
-  }
-
-  text(): string {
-    return this._text;
-  }
-
-  args(): JSHandle[] {
-    return this._args;
-  }
-
-  location(): ConsoleMessageLocation {
-    return this._location;
-  }
-}
-
-export class FileChooser {
-  _client: CDPSession;
-  _element: ElementHandle;
-  _multiple: boolean;
-  _handled = false;
-
-  constructor(
-    client: CDPSession,
-    element: ElementHandle,
-    event: Protocol.Page.fileChooserOpenedPayload
-  ) {
-    this._client = client;
-    this._element = element;
-    this._multiple = event.mode !== 'selectSingle';
-  }
-
-  isMultiple(): boolean {
-    return this._multiple;
-  }
-
-  async accept(filePaths: string[]): Promise<void> {
-    assert(
-      !this._handled,
-      'Cannot accept FileChooser which is already handled!'
-    );
-    this._handled = true;
-    await this._element.uploadFile(...filePaths);
-  }
-
-  async cancel(): Promise<void> {
-    assert(
-      !this._handled,
-      'Cannot cancel FileChooser which is already handled!'
-    );
-    this._handled = true;
-  }
 }
